requirements.txt
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
distro==1.6.0
Django==3.2.12
django-dotenv==1.4.2
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
paramiko==2.9.2
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
websocket-client==0.59.0
zipp==3.6.0
